requirements.txt
groq>=0.9.0
langgraph>=1.1.0
langchain-core>=0.3.0
openpyxl>=3.1.0
pandas>=2.0.0
numpy>=1.26.0
sentence-transformers>=5.0.0
scikit-learn>=1.4.0
matplotlib>=3.8.0
seaborn>=0.13.0
python-dotenv>=1.0.0
tqdm>=4.66.0
pydantic>=2.0.0
torch>=2.0.0
